% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_ridge.R
\name{apply_na_mask_and_weights}
\alias{apply_na_mask_and_weights}
\title{Apply NA Mask and Weights}
\usage{
apply_na_mask_and_weights(Y, X, na_mask = NULL, weights = NULL)
}
\description{
Removes rows indicated by `na_mask` and subsets `weights` accordingly.
}
\keyword{internal}
