% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_gdlite.R
\name{calculate_noise_pool_mask}
\alias{calculate_noise_pool_mask}
\title{Determine the noise pool mask for GLMdenoise-Lite}
\usage{
calculate_noise_pool_mask(
  r2_cv_initial,
  tsnr_values,
  r2_thresh_noise_pool,
  tsnr_thresh_noise_pool,
  is_single_run
)
}
\arguments{
\item{r2_cv_initial}{Initial leave-one-run-out R-squared values.}

\item{tsnr_values}{Temporal SNR values.}

\item{r2_thresh_noise_pool}{R-squared threshold for defining the noise pool.}

\item{tsnr_thresh_noise_pool}{tSNR threshold for defining the noise pool.}

\item{is_single_run}{Logical flag indicating whether CV failed or only one run.}
}
\value{
Logical vector marking voxels in the noise pool.
}
\description{
Determine the noise pool mask for GLMdenoise-Lite
}
\keyword{internal}
