% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_utils.R
\name{calculate_R2_voxelwise}
\alias{calculate_R2_voxelwise}
\title{Calculate Voxel-wise R-squared}
\usage{
calculate_R2_voxelwise(Y_observed, Y_residuals)
}
\arguments{
\item{Y_observed}{Matrix of observed data (timepoints x voxels).}

\item{Y_residuals}{Matrix of residuals from a model (timepoints x voxels).}
}
\value{
A numeric vector of R-squared values, one for each voxel.
}
\description{
Calculate Voxel-wise R-squared
}
\keyword{internal}
