% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_utils.R
\name{calculate_residuals_ols}
\alias{calculate_residuals_ols}
\title{Calculate OLS residuals using lm.fit}
\usage{
calculate_residuals_ols(Y, X)
}
\arguments{
\item{Y}{Dependent variable matrix (timepoints x variables/voxels).}

\item{X}{Design matrix (timepoints x regressors).}
}
\value{
Matrix of residuals (timepoints x variables/voxels).
}
\description{
Calculate OLS residuals using lm.fit
}
\keyword{internal}
