% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_ridge.R
\name{construct_penalty}
\alias{construct_penalty}
\title{Construct Ridge Penalty}
\usage{
construct_penalty(
  X_eff,
  Y_eff,
  K_penalty_diag = NULL,
  projection_mats = NULL,
  lambda_values = NULL,
  gcv_lambda = FALSE,
  res_var_scale = 1,
  lambda_grid = 10^seq(-2, 2, length.out = 5),
  K_penalty_mat = NULL,
  use_penalty_matrix = FALSE
)
}
\description{
Builds the penalty matrix or diagonal from projection matrices and lambda values.
}
\keyword{internal}
