% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_progressive_viz.R
\name{create_beta_stability_plot}
\alias{create_beta_stability_plot}
\title{Create Beta Stability Plot}
\usage{
create_beta_stability_plot(workflow_output, stages)
}
\arguments{
\item{workflow_output}{Output from `NDX_Process_Subject`. Must contain a
`beta_history_per_pass` list with beta matrices for each pass.}

\item{stages}{Vector of stage names}
}
\value{
Plotly plot object
}
\description{
Create Beta Stability Plot
}
\keyword{internal}
