% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_progressive_viz.R
\name{create_des_progression_plot}
\alias{create_des_progression_plot}
\title{Create DES Progression Plot}
\usage{
create_des_progression_plot(workflow_output, stages)
}
\arguments{
\item{workflow_output}{Output from `NDX_Process_Subject`. Must contain a
`diagnostics_per_pass` list with a numeric `DES` element for each pass.}

\item{stages}{Vector of stage names}
}
\value{
Plotly plot object
}
\description{
Create DES Progression Plot
}
\keyword{internal}
