% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_progressive_viz.R
\name{create_psd_progression_plot}
\alias{create_psd_progression_plot}
\title{Create PSD Progression Plot}
\usage{
create_psd_progression_plot(workflow_output, stages)
}
\arguments{
\item{workflow_output}{Output from `NDX_Process_Subject`. Must contain
`pass0_residuals` and `Y_residuals_final_unwhitened` for PSD estimation.}

\item{stages}{Vector of stage names}
}
\value{
Plotly plot object
}
\description{
Create PSD Progression Plot
}
\keyword{internal}
