% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_hrf.R
\name{cv_fusedlasso}
\alias{cv_fusedlasso}
\title{Block-wise K-fold Cross-Validation for Fused Lasso (glmgen)}
\usage{
cv_fusedlasso(y, X, lambda_grid, gamma_grid, k_folds, block_ids)
}
\description{
Block-wise K-fold Cross-Validation for Fused Lasso (glmgen)
}
\keyword{internal}
