% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_gdlite.R
\name{define_good_voxel_mask}
\alias{define_good_voxel_mask}
\title{Define the good voxel mask used for K selection}
\usage{
define_good_voxel_mask(
  r2_cv_initial,
  tsnr_values,
  r2_thresh_good_voxels,
  is_single_run,
  n_voxels
)
}
\arguments{
\item{r2_cv_initial}{Initial cross-validated R-squared values.}

\item{tsnr_values}{tSNR values.}

\item{r2_thresh_good_voxels}{Threshold for good voxels.}

\item{is_single_run}{Logical indicating single-run or failed CV.}

\item{n_voxels}{Total number of voxels.}
}
\value{
Logical vector defining the good voxel mask.
}
\description{
Define the good voxel mask used for K selection
}
\keyword{internal}
