% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_spectral.R
\name{.empty_spec_matrix}
\alias{.empty_spec_matrix}
\title{Create an empty matrix for spectral regressors}
\usage{
.empty_spec_matrix(n_rows)
}
\description{
Create an empty matrix for spectral regressors
}
\keyword{internal}
