% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann_merge_cpp.R
\name{.grassmann_merge_iterative_rcpp}
\alias{.grassmann_merge_iterative_rcpp}
\title{Rcpp implementation of iterative Grassmann merge}
\usage{
.grassmann_merge_iterative_rcpp(V_list_valid, k_target_global)
}
\arguments{
\item{V_list_valid}{List of matrices (voxels x k_r).}

\item{k_target_global}{Target number of global components.}
}
\value{
Matrix V_global with attribute `svd_times`.
}
\description{
Rcpp implementation of iterative Grassmann merge
}
\keyword{internal}
