% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_add_annihilation_results}
\alias{.ndx_add_annihilation_results}
\title{Add annihilation mode specific results}
\usage{
.ndx_add_annihilation_results(final_results, workflow_state)
}
\description{
Add annihilation mode specific results
}
\keyword{internal}
