% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step5_design_matrix.R
\name{.ndx_add_orthogonalized_components}
\alias{.ndx_add_orthogonalized_components}
\title{Add orthogonalized components to design matrix}
\usage{
.ndx_add_orthogonalized_components(X_full_design, workflow_state, verbose)
}
\description{
Add orthogonalized components to design matrix
}
\keyword{internal}
