% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step5_design_matrix.R
\name{.ndx_build_annihilation_design_matrix}
\alias{.ndx_build_annihilation_design_matrix}
\title{Build design matrix for annihilation mode}
\usage{
.ndx_build_annihilation_design_matrix(
  workflow_state,
  current_pass_results,
  verbose
)
}
\description{
Build design matrix for annihilation mode
}
\keyword{internal}
