% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_build_projection_matrices}
\alias{.ndx_build_projection_matrices}
\title{Build projection matrices for anisotropic ridge}
\usage{
.ndx_build_projection_matrices(regressor_indices, n_regressors, opts_ridge)
}
\description{
Build projection matrices for anisotropic ridge
}
\keyword{internal}
