% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step5_design_matrix.R
\name{.ndx_build_standard_design_matrix}
\alias{.ndx_build_standard_design_matrix}
\title{Build standard design matrix}
\usage{
.ndx_build_standard_design_matrix(workflow_state, current_pass_results)
}
\description{
Build standard design matrix
}
\keyword{internal}
