% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_calculate_ljung_box}
\alias{.ndx_calculate_ljung_box}
\title{Calculate Ljung-Box test statistic}
\usage{
.ndx_calculate_ljung_box(current_pass_results)
}
\description{
Calculate Ljung-Box test statistic
}
\keyword{internal}
