% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_calculate_pass_residuals}
\alias{.ndx_calculate_pass_residuals}
\title{Calculate residuals for next pass}
\usage{
.ndx_calculate_pass_residuals(workflow_state, current_pass_results)
}
\description{
Calculate residuals for next pass
}
\keyword{internal}
