% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_calculate_rho_noise_projection}
\alias{.ndx_calculate_rho_noise_projection}
\title{Calculate Rho Noise Projection and handle annihilation mode}
\usage{
.ndx_calculate_rho_noise_projection(
  workflow_state,
  current_pass_results,
  verbose
)
}
\description{
Calculate Rho Noise Projection and handle annihilation mode
}
\keyword{internal}
