% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_check_convergence}
\alias{.ndx_check_convergence}
\title{Check convergence criteria}
\usage{
.ndx_check_convergence(workflow_state, pass_num)
}
\description{
Check convergence criteria
}
\keyword{internal}
