% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_compute_rho_projection}
\alias{.ndx_compute_rho_projection}
\title{Compute rho noise projection}
\usage{
.ndx_compute_rho_projection(
  U_NDX_Nuisance_Combined_list,
  workflow_state,
  verbose
)
}
\description{
Compute rho noise projection
}
\keyword{internal}
