% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step3_rpca.R
\name{.ndx_determine_rpca_k}
\alias{.ndx_determine_rpca_k}
\title{Determine RPCA k value}
\usage{
.ndx_determine_rpca_k(workflow_state, pass_num, current_opts_rpca)
}
\description{
Determine RPCA k value
}
\keyword{internal}
