% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_finalize_workflow_results}
\alias{.ndx_finalize_workflow_results}
\title{Finalize workflow results}
\usage{
.ndx_finalize_workflow_results(workflow_state)
}
\description{
Finalize workflow results
}
\keyword{internal}
