% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_rpca_helpers.R
\name{.ndx_form_temporal_components}
\alias{.ndx_form_temporal_components}
\title{Form Temporal Components From Residuals}
\usage{
.ndx_form_temporal_components(residuals_list, V_global)
}
\arguments{
\item{residuals_list}{List of residual matrices (Time x Voxels) for each run.}

\item{V_global}{Voxel-space basis matrix returned from merging.}
}
\value{
Matrix of concatenated temporal components.
}
\description{
Multiplies each run's residuals by the global voxel-space basis and
concatenates the results.
}
\keyword{internal}
