% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_identify_regressor_types}
\alias{.ndx_identify_regressor_types}
\title{Identify regressor types for anisotropic ridge}
\usage{
.ndx_identify_regressor_types(regressor_names)
}
\description{
Identify regressor types for anisotropic ridge
}
\keyword{internal}
