% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_initialize_workflow_state}
\alias{.ndx_initialize_workflow_state}
\title{Initialize workflow state for NDX processing}
\usage{
.ndx_initialize_workflow_state(
  Y_fmri,
  events,
  motion_params,
  run_idx,
  TR,
  spike_TR_mask,
  user_options,
  verbose
)
}
\description{
Initialize workflow state for NDX processing
}
\keyword{internal}
