% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_orthogonalize_annihilation_components}
\alias{.ndx_orthogonalize_annihilation_components}
\title{Orthogonalize components for annihilation mode}
\usage{
.ndx_orthogonalize_annihilation_components(
  U_NDX_Nuisance_Combined_list,
  U_GD_Lite_fixed_PCs,
  verbose_workflow
)
}
\description{
Orthogonalize components for annihilation mode
}
\keyword{internal}
