% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step3_rpca.R
\name{.ndx_prepare_adaptive_rpca_options}
\alias{.ndx_prepare_adaptive_rpca_options}
\title{Prepare adaptive RPCA options}
\usage{
.ndx_prepare_adaptive_rpca_options(workflow_state, pass_num, opts_rpca)
}
\description{
Prepare adaptive RPCA options
}
\keyword{internal}
