% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step2_hrf.R
\name{.ndx_process_hrfs_for_design}
\alias{.ndx_process_hrfs_for_design}
\title{Process HRFs for design matrix construction}
\usage{
.ndx_process_hrfs_for_design(estimated_hrfs_raw, events, pass_num, verbose)
}
\description{
Process HRFs for design matrix construction
}
\keyword{internal}
