% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step3_rpca.R
\name{.ndx_process_rpca_results}
\alias{.ndx_process_rpca_results}
\title{Process RPCA results}
\usage{
.ndx_process_rpca_results(
  workflow_state,
  rpca_out,
  current_pass_results,
  verbose
)
}
\description{
Process RPCA results
}
\keyword{internal}
