% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_run_anisotropic_ridge}
\alias{.ndx_run_anisotropic_ridge}
\title{Run anisotropic ridge regression}
\usage{
.ndx_run_anisotropic_ridge(
  current_pass_results,
  opts_ridge,
  precision_weights_for_pass,
  verbose
)
}
\description{
Run anisotropic ridge regression
}
\keyword{internal}
