% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_run_isotropic_ridge}
\alias{.ndx_run_isotropic_ridge}
\title{Run isotropic ridge regression}
\usage{
.ndx_run_isotropic_ridge(
  current_pass_results,
  opts_ridge,
  precision_weights_for_pass,
  n_regressors
)
}
\description{
Run isotropic ridge regression
}
\keyword{internal}
