% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_run_iterative_passes}
\alias{.ndx_run_iterative_passes}
\title{Run iterative refinement passes}
\usage{
.ndx_run_iterative_passes(workflow_state)
}
\description{
Run iterative refinement passes
}
\keyword{internal}
