% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_run_ridge_regression}
\alias{.ndx_run_ridge_regression}
\title{Run ridge regression with anisotropic or isotropic penalties}
\usage{
.ndx_run_ridge_regression(current_pass_results, opts_ridge, verbose)
}
\description{
Run ridge regression with anisotropic or isotropic penalties
}
\keyword{internal}
