% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow.R
\name{.ndx_run_single_pass}
\alias{.ndx_run_single_pass}
\title{Run a single pass of the NDX workflow}
\usage{
.ndx_run_single_pass(workflow_state, pass_num)
}
\description{
Run a single pass of the NDX workflow
}
\keyword{internal}
