% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step5_design_matrix.R
\name{.ndx_step_design_matrix}
\alias{.ndx_step_design_matrix}
\title{Step 5: Construct Design Matrix}
\usage{
.ndx_step_design_matrix(workflow_state, pass_num, current_pass_results)
}
\description{
Step 5: Construct Design Matrix
}
\keyword{internal}
