% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_step_diagnostics}
\alias{.ndx_step_diagnostics}
\title{Step 8: Calculate Diagnostics}
\usage{
.ndx_step_diagnostics(workflow_state, pass_num, current_pass_results)
}
\description{
Step 8: Calculate Diagnostics
}
\keyword{internal}
