% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step2_hrf.R
\name{.ndx_step_hrf_estimation}
\alias{.ndx_step_hrf_estimation}
\title{Step 2: HRF Estimation}
\usage{
.ndx_step_hrf_estimation(workflow_state, pass_num, current_pass_results)
}
\description{
Step 2: HRF Estimation
}
\keyword{internal}
