% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step1_initial_glm.R
\name{.ndx_step_initial_glm}
\alias{.ndx_step_initial_glm}
\title{Step 1: Initial GLM or use previous residuals}
\usage{
.ndx_step_initial_glm(workflow_state, pass_num, current_pass_results)
}
\description{
Step 1: Initial GLM or use previous residuals
}
\keyword{internal}
