% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_step_ridge_regression}
\alias{.ndx_step_ridge_regression}
\title{Step 7: Ridge Regression}
\usage{
.ndx_step_ridge_regression(workflow_state, pass_num, current_pass_results)
}
\description{
Step 7: Ridge Regression
}
\keyword{internal}
