% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step3_rpca.R
\name{.ndx_step_rpca_components}
\alias{.ndx_step_rpca_components}
\title{Step 3: RPCA Nuisance Components}
\usage{
.ndx_step_rpca_components(workflow_state, pass_num, current_pass_results)
}
\description{
Step 3: RPCA Nuisance Components
}
\keyword{internal}
