% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step4_spectral.R
\name{.ndx_step_spectral_components}
\alias{.ndx_step_spectral_components}
\title{Step 4: Spectral Nuisance Components}
\usage{
.ndx_step_spectral_components(workflow_state, pass_num, current_pass_results)
}
\description{
Step 4: Spectral Nuisance Components
}
\keyword{internal}
