% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step6_whitening.R
\name{.ndx_step_whitening}
\alias{.ndx_step_whitening}
\title{Step 6: AR(2) Pre-whitening}
\usage{
.ndx_step_whitening(workflow_state, pass_num, current_pass_results)
}
\description{
Step 6: AR(2) Pre-whitening
}
\keyword{internal}
