% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step8_diagnostics.R
\name{.ndx_store_additional_diagnostics}
\alias{.ndx_store_additional_diagnostics}
\title{Store additional diagnostics}
\usage{
.ndx_store_additional_diagnostics(
  pass_diagnostics,
  current_pass_results,
  annihilation_active_this_pass
)
}
\description{
Store additional diagnostics
}
\keyword{internal}
