% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step7_ridge.R
\name{.ndx_tune_lambda_parameters}
\alias{.ndx_tune_lambda_parameters}
\title{Tune lambda parameters for anisotropic ridge}
\usage{
.ndx_tune_lambda_parameters(
  current_pass_results,
  proj_mats,
  opts_ridge,
  res_var_est
)
}
\description{
Tune lambda parameters for anisotropic ridge
}
\keyword{internal}
