% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_hrf.R
\name{.perform_hrf_clustering}
\alias{.perform_hrf_clustering}
\title{Perform K-Medoids Clustering on Voxel Time Courses}
\usage{
.perform_hrf_clustering(Y_for_clustering, num_clusters, user_options, verbose)
}
\description{
Perform K-Medoids Clustering on Voxel Time Courses
}
\keyword{internal}
