% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_hrf.R
\name{get_fir_design_matrix_for_condition}
\alias{get_fir_design_matrix_for_condition}
\title{Generate FIR Design Matrix for a Specific Condition}
\usage{
get_fir_design_matrix_for_condition(
  condition_name,
  events_df,
  sampling_frame,
  fir_taps,
  TR
)
}
\description{
Generate FIR Design Matrix for a Specific Condition
}
\keyword{internal}
