% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_utils.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge Two Lists with Defaults}
\usage{
merge_lists(defaults, user)
}
\arguments{
\item{defaults}{A list of default values.}

\item{user}{User-supplied options.}
}
\value{
Combined list.
}
\description{
Wrapper around `utils::modifyList` that also handles `NULL` user lists.
}
