% Generated by hand for tests
% Please edit documentation in R/ndx_ridge.R
\name{ndx_compute_projection_matrices}
\alias{ndx_compute_projection_matrices}
\title{Compute Projection Matrices for Anisotropic Ridge}
\usage{
ndx_compute_projection_matrices(U_GD = NULL, U_Unique = NULL,
  U_Noise = NULL, n_regressors)
}
\arguments{
\item{U_GD}{Matrix of GLMdenoise PCs.}
\item{U_Unique}{Matrix of unique nuisance components.}
\item{U_Noise}{Matrix of noise components.}
\item{n_regressors}{Total number of regressors.}
}
\value{List of projection matrices.}
\description{Helper used by anisotropic ridge regression to build projectors.}
