% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_diagnostic_plots.R
\name{ndx_plot_des_per_pass}
\alias{ndx_plot_des_per_pass}
\title{Plot DES per pass and save to PNG}
\usage{
ndx_plot_des_per_pass(diagnostics_per_pass, output_dir)
}
\description{
Plot DES per pass and save to PNG
}
\keyword{internal}
