% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_diagnostic_plots.R
\name{ndx_plot_gdlite_r2_by_k}
\alias{ndx_plot_gdlite_r2_by_k}
\title{Plot GLMdenoise R2 by k}
\usage{
ndx_plot_gdlite_r2_by_k(gdlite_diagnostics, output_dir)
}
\description{
Plot GLMdenoise R2 by k
}
\keyword{internal}
