% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_diagnostic_plots.R
\name{ndx_plot_ljung_box_pvalues}
\alias{ndx_plot_ljung_box_pvalues}
\title{Plot Ljung-Box p-values per pass}
\usage{
ndx_plot_ljung_box_pvalues(diagnostics_per_pass, output_dir)
}
\description{
Plot Ljung-Box p-values per pass
}
\keyword{internal}
