% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_diagnostic_plots.R
\name{ndx_plot_spike_carpet}
\alias{ndx_plot_spike_carpet}
\title{Plot spike carpet from RPCA S matrix}
\usage{
ndx_plot_spike_carpet(S_matrix, output_dir)
}
\description{
Plot spike carpet from RPCA S matrix
}
\keyword{internal}
