% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_workflow_utils.R
\name{ndx_prepare_workflow_options}
\alias{ndx_prepare_workflow_options}
\title{Merge user options with defaults}
\usage{
ndx_prepare_workflow_options(user_options = list())
}
\arguments{
\item{user_options}{List of user supplied options.}
}
\value{
Named list of options where unspecified values are filled with defaults.
}
\description{
Combines user-supplied workflow options with the ND-X defaults,
filling in any unspecified values.
}
