% Generated by hand for tests
% Please edit documentation in R/ndx_ridge.R
\name{ndx_update_lambda_aggressiveness}
\alias{ndx_update_lambda_aggressiveness}
\title{Adjust Lambda Based on Residual Projection}
\usage{
ndx_update_lambda_aggressiveness(lambda_parallel, rho_noise_projection,
  high_threshold = 0.1, low_threshold = 0.03,
  increase_factor = 1.25, decrease_factor = 0.8)
}
\arguments{
\item{lambda_parallel}{Current lambda value.}
\item{rho_noise_projection}{Proportion of residual variance in noise subspace.}
\item{high_threshold}{Increase threshold.}
\item{low_threshold}{Decrease threshold.}
\item{increase_factor}{Factor by which to increase.}
\item{decrease_factor}{Factor by which to decrease.}
}
\value{Updated lambda value.}
\description{Heuristic adjustment of lambda based on residual noise.}
