% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_options.R
\name{ndx_user_options}
\alias{ndx_user_options}
\title{ND-X User Options Reference}
\description{
This documentation describes all configuration options available in the
ND-X package functions. Most functions accept a \code{user_options} parameter,
which is a list containing these configuration options organized into
sub-lists for different modules.
}
NULL
