% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_utils.R
\name{ndx_val_or_na}
\alias{ndx_val_or_na}
\title{Return `NA` if Value is `NULL`}
\usage{
ndx_val_or_na(x)
}
\arguments{
\item{x}{An object that might be `NULL`.}
}
\value{
`NA` if `x` is `NULL`, otherwise `x`.
}
\description{
Simple helper that returns `NA` when the input is `NULL`, otherwise the
input value unchanged.
}
