% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_hrf.R
\name{project_cone_heuristic}
\alias{project_cone_heuristic}
\title{Basic Cone Projection Heuristic for FIR Coefficients}
\usage{
project_cone_heuristic(fir_coeffs)
}
\description{
Basic Cone Projection Heuristic for FIR Coefficients
}
\keyword{internal}
