% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_gdlite.R
\name{run_final_glm_gdlite}
\alias{run_final_glm_gdlite}
\title{Run the final GLM for GLMdenoise-Lite}
\usage{
run_final_glm_gdlite(Y_fmri, X_gd, selected_pcs, K_star)
}
\arguments{
\item{Y_fmri}{fMRI data matrix.}

\item{X_gd}{Base design matrix.}

\item{selected_pcs}{Matrix of selected principal components.}

\item{K_star}{Number of selected PCs.}
}
\value{
List with `betas` and `residuals` from the final GLM.
}
\description{
Run the final GLM for GLMdenoise-Lite
}
\keyword{internal}
