% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_hrf.R
\name{validate_hrf_inputs}
\alias{validate_hrf_inputs}
\title{Validate Inputs for HRF Estimation}
\usage{
validate_hrf_inputs(
  Y_fmri,
  pass0_residuals,
  events,
  run_idx,
  TR,
  spike_TR_mask,
  user_options
)
}
\arguments{
\item{Y_fmri}{Matrix of fMRI data.}

\item{pass0_residuals}{Matrix of residuals.}

\item{events}{Data frame of experimental events.}

\item{run_idx}{Numeric vector of run membership.}

\item{TR}{Numeric, repetition time.}

\item{spike_TR_mask}{Optional logical vector for spike TRs.}

\item{user_options}{A list of user-configurable options.}
}
\value{
A list containing validated and potentially augmented `user_options`,
  and `n_timepoints` and `validated_spike_TR_mask`.
}
\description{
Helper function to validate inputs for `ndx_estimate_initial_hrfs`.
Also merges user_options with defaults.
}
\keyword{internal}
