% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndx_utils.R
\name{calculate_R2_voxelwise}
\alias{calculate_R2_voxelwise}
\title{Calculate Voxel-wise R-squared}
\usage{
calculate_R2_voxelwise(Y_observed, Y_residuals)
}
\arguments{
\item{Y_observed}{Matrix of observed data (timepoints x voxels).}

\item{Y_residuals}{Matrix of residuals from a model (timepoints x voxels).}
}
\value{
A numeric vector of R-squared values, one for each voxel.
  Names are stripped from the returned vector.
}
\description{
Calculate Voxel-wise R-squared
}
\details{
Valid observations for each voxel are selected using
  `!is.na(Y_observed[, j])`. Total Sum of Squares (TSS) and Residual Sum of
  Squares (RSS) are computed from these valid values only. If the calculated
  TSS is zero the returned R\eqn{^2} for that voxel is set to zero.
}
